Atividade 6 e 8/libraryon_api/requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.1
cycler==0.10.0
defusedxml==0.5.0
Django==2.1
django-braces==1.13.0
django-cors-middleware==1.3.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-filters==0.2.1
django-guardian==1.4.9
django-nose==1.4.5
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==3.2.3
httplib2==0.11.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
model-mommy==1.5.1
nose==1.3.7
numpy==1.14.3
oauthlib==2.1.0
openapi-codec==1.3.2
Pillow==5.2.0
psycopg2==2.7.5
Pygments==2.2.0
PyJWT==1.6.4
pyparsing==2.2.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==1.0.0
simplejson==3.15.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
uritemplate==3.0.0
urllib3==1.22
